random/my-app/app/backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
boto3==1.34.149
botocore==1.34.149
certifi==2024.2.2
distro==1.9.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
openai==0.28.0
pydantic==2.7.2
pydantic_core==2.18.3
python-dateutil==2.9.0.post0
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==1.26.18
xmltodict==0.13.0
celery[sqs]==5.3.1
anthropic==0.30.1
tesseract==0.1.3
pytesseract==0.3.10
Pillow==10.1.0
awslambdaric

